refactor: improved url validation on some modules
Upcoming change would break the validations for these modules if these changes aren't made. This is due to AniList redirecting media that are missing their slug.

Example `/anime/1` redirects to `/anime/1/Cowboy-Bebop/`.

This would cause the modules to sometimes load twice. Would be better to fix it where the module is validated/loaded but for now this'll do.